+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - dotfiles now only apply at `run`-time. Updating dotfiles does not lead to a re-build
 - Add `DOTFILES_PATH` to `run.conf` to let user configure where dotfiles are
+- base "dotfiles" can be seeded from e.g. `XDG_CONFIG_HOME/paul-envs/dotfiles`
 
 ### Bug fixes